+++ R/PLS_beta.R
@@ -119,7 +119,7 @@
 #' @return Depends on the model that was used to fit the model.
 #' @note Use \code{plsRbeta} instead.
 #' @author Frédéric Bertrand\cr
-#' \email{frederic.bertrand@@utt.fr}\cr
+#' \email{frederic.bertrand@@lecnam.net}\cr
 #' \url{https://fbertran.github.io/homepage/}
 #' @seealso \code{\link[plsRbeta]{PLS_beta_wvc}} and
 #' \code{\link[plsRbeta]{PLS_beta_kfoldcv}}
@@ -128,7 +128,7 @@
 #' Myriam Maumy-Bertrand (2013). Régression Bêta
 #' PLS. \emph{Journal de la Société Française de Statistique},
 #' \bold{154}(3):143-159.
-#' \url{http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/215}
+#' \url{https://ojs-test.apps.ocp.math.cnrs.fr/index.php/J-SFdS/article/view/215}
 #' @keywords models regression
 #' @examples
 #'