+++ R/bootplsbeta.R
@@ -40,7 +40,7 @@
 #' @return An object of class \code{"boot"}. See the Value part of the help of
 #' the function \code{\link[boot:boot]{boot}}.
 #' @author Frédéric Bertrand\cr
-#' \email{frederic.bertrand@@utt.fr}\cr
+#' \email{frederic.bertrand@@lecnam.net}\cr
 #' \url{https://fbertran.github.io/homepage/}
 #' @seealso \code{\link[boot:boot]{boot}}
 #' @references Frédéric Bertrand, Nicolas Meyer,
@@ -48,7 +48,7 @@
 #' Myriam Maumy-Bertrand (2013). Régression Bêta
 #' PLS. \emph{Journal de la Société Française de Statistique},
 #' \bold{154}(3):143-159.
-#' \url{http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/215}
+#' \url{https://ojs-test.apps.ocp.math.cnrs.fr/index.php/J-SFdS/article/view/215}
 #' @keywords models
 #' @examples
 #' \donttest{