+++ R/permcoefs.plsRbeta.raw.R
@@ -22,7 +22,7 @@
 #' @param verbose should info messages be displayed ?
 #' @return Estimates on a bootstrap sample.
 #' @author Frédéric Bertrand\cr
-#' \email{frederic.bertrand@@utt.fr}\cr
+#' \email{frederic.bertrand@@lecnam.net}\cr
 #' \url{https://fbertran.github.io/homepage/}
 #' @seealso See also \code{\link{bootplsbeta}}.
 #' @references Frédéric Bertrand, Nicolas Meyer,
@@ -30,7 +30,7 @@
 #' Myriam Maumy-Bertrand (2013). Régression Bêta
 #' PLS. \emph{Journal de la Société Française de Statistique},
 #' \bold{154}(3):143-159.
-#' \url{http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/215}
+#' \url{https://ojs-test.apps.ocp.math.cnrs.fr/index.php/J-SFdS/article/view/215}
 #' @keywords models
 #' @examples
 #' \donttest{