+++ R/plsRbeta-package.R
@@ -1,34 +1,6 @@
-#' @title plsRbeta-package
+#' @keywords internal
+#' @aliases plsRbeta-package NULL
 #'
-#' @description Provides Partial least squares Regression for (weighted) beta regression models (Bertrand 2013,  <http://journal-sfds.fr/article/view/215>) and k-fold cross-validation of such models using various criteria. It allows for missing data in the explanatory variables. Bootstrap confidence intervals constructions are also available.
+#' @references Partial least squares Regression for (weighted) beta regression models (Bertrand 2013,  <https://ojs-test.apps.ocp.math.cnrs.fr/index.php/J-SFdS/article/view/215>), \url{https://github.com/fbertran/plsRbeta/} et \url{https://fbertran.github.io/plsRbeta/}
 #'
-#' @docType package
-#' @name plsRbeta-package
-#' @references Partial least squares Regression for (weighted) beta regression models (Bertrand 2013,  <http://journal-sfds.fr/article/view/215>), \url{https://github.com/fbertran/plsRbeta/} et \url{https://fbertran.github.io/plsRbeta/}
-#' 
-#' @examples
-#' 
-#' data("GasolineYield",package="betareg")
-#' modpls <- plsRbeta(yield~.,data=GasolineYield,nt=3,modele="pls-beta")
-#' modpls$pp
-#' modpls$Coeffs
-#' modpls$Std.Coeffs
-#' modpls$InfCrit
-#' modpls$PredictY[1,]
-#' rm("modpls")
-#' 
-#' data("GasolineYield",package="betareg")
-#' yGasolineYield <- GasolineYield$yield
-#' XGasolineYield <- GasolineYield[,2:5]
-#' modpls <- plsRbeta(yGasolineYield,XGasolineYield,nt=3,modele="pls-beta")
-#' modpls$pp
-#' modpls$Coeffs
-#' modpls$Std.Coeffs
-#' modpls$InfCrit
-#' modpls$PredictY[1,]
-#' rm("modpls")
-#' 
-NULL
-
-
-
+"_PACKAGE"