+++ R/print.summary.plsRbetamodel.R
@@ -8,7 +8,7 @@
 #' @param \dots not used
 #' @return \item{language}{call of the model}
 #' @author Frédéric Bertrand\cr
-#' \email{frederic.bertrand@@utt.fr}\cr
+#' \email{frederic.bertrand@@lecnam.net}\cr
 #' \url{https://fbertran.github.io/homepage/}
 #' @seealso \code{\link{print}} and \code{\link{summary}}
 #' @references Frédéric Bertrand, Nicolas Meyer,
@@ -16,7 +16,7 @@
 #' Myriam Maumy-Bertrand (2013). Régression Bêta
 #' PLS. \emph{Journal de la Société Française de Statistique},
 #' \bold{154}(3):143-159.
-#' \url{http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/215}
+#' \url{https://ojs-test.apps.ocp.math.cnrs.fr/index.php/J-SFdS/article/view/215}
 #' @keywords methods print
 #' @examples
 #'