+++ R/simul_data_UniYX_beta.R
@@ -21,7 +21,7 @@
 #' @param phi0 Simulation scheme "a" parameter
 #' @return \item{vector}{\eqn{(Y,X_1,\ldots,X_{totdim})}}
 #' @author Frédéric Bertrand\cr
-#' \email{frederic.bertrand@@utt.fr}\cr
+#' \email{frederic.bertrand@@lecnam.net}\cr
 #' \url{https://fbertran.github.io/homepage/}
 #' @seealso \code{\link[plsRglm]{simul_data_UniYX}}
 #' @references Frédéric Bertrand, Nicolas Meyer,
@@ -29,7 +29,7 @@
 #' Myriam Maumy-Bertrand (2013). Régression Bêta
 #' PLS. \emph{Journal de la Société Française de Statistique},
 #' \bold{154}(3):143-159.
-#' \url{http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/215}
+#' \url{https://ojs-test.apps.ocp.math.cnrs.fr/index.php/J-SFdS/article/view/215}
 #' 
 #' T. Naes, H. Martens (1985). Comparison of prediction methods for
 #' multicollinear data.  \emph{Commun. Stat., Simul.}, \bold{14}:545-576.