+++ tests/testthat.R
@@ -1,0 +1,3 @@
+library(testthat)
+library(plsRbeta)
+test_check("plsRbeta")